Правки по запуску и удалению пароля

+++ noPass.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 
-.stop.sh
-rm nginx/.htpasswd
-.start.sh
+if [ -f "nginx/.htpasswd" ]; then
+    rm nginx/.htpasswd
+    echo "✓ Пароль для web приложения удалён"
+else
+    echo "✓ Пароль для web приложения уже удалён"
+fi
 
-echo "✓ Пароль для web приложения удалён"
+# grep -q "mo-cctv-" вернёт 0 при `echo $?` если найдено и 1 если не найдено
+if docker ps --filter "name=mo-cctv-" --format "table {{.Names}}" | grep -q "mo-cctv-"; then
+    ./restart.sh
+fi
